Fix global stats generation
